+++ include/Game/THPPlayer.h
@@ -58,8 +58,8 @@ struct THPPlayer : public JKRDisposer, public CNode {
 	// vtable 2 (CNode + self)
 	virtual void update();                              // _1C
 	virtual void draw(Graphics&);                       // _20
-	virtual void draw(Graphics&, long, long, u32, u32); // _24
-	virtual void draw(Graphics&, long, long, f32);      // _28
+	virtual void draw(Graphics&, s32, s32, u32, u32); // _24
+	virtual void draw(Graphics&, s32, s32, f32);      // _28
 
 	void load(EMovieIndex);
 	void load();
