+++ include/JSystem/JKernel/JKRFileLoader.h
@@ -20,14 +20,14 @@ struct JKRFileLoader : public JKRDisposer {
 	virtual void removeResourceAll()                                                              = 0; // _24
 	virtual bool removeResource(void*)                                                            = 0; // _28
 	virtual bool detachResource(void*)                                                            = 0; // _2C
-	virtual long getResSize(const void*) const                                                    = 0; // _30
+	virtual s32 getResSize(const void*) const                                                    = 0; // _30
 	virtual u32 countFile(const char*) const                                                      = 0; // _34
 	virtual JKRFileFinder* getFirstFile(const char*) const                                        = 0; // _38
 
 	static void* getGlbResource(char const*);
 	static void* getGlbResource(char const*, JKRFileLoader*);
 	static JKRFileLoader* findVolume(char const**);
-	static const char* fetchVolumeName(char*, long, char const*);
+	static const char* fetchVolumeName(char*, s32, char const*);
 
 	bool isMounted() const { return mIsMounted; }
 	u32 getVolumeType() const { return mMagicWord; }
