+++ include/PowerPC_EABI_Support/Runtime/NMWException.h
@@ -9,7 +9,7 @@
 extern "C" {
 #endif
 
-typedef short vbase_ctor_arg_type;
+typedef s16 vbase_ctor_arg_type;
 typedef char local_cond_type;
 
 typedef struct CatchInfo {
@@ -17,7 +17,7 @@ typedef struct CatchInfo {
 	void* typeinfo;
 	void* dtor;
 	void* sublocation;
-	long pointercopy;
+	s32 pointercopy;
 	void* stacktop;
 } CatchInfo;
 
@@ -32,7 +32,7 @@ extern void __destroy_global_chain(void);
 
 extern void __end__catch(CatchInfo* catchinfo);
 extern void __throw(char* throwtype, void* location, void* dtor);
-extern char __throw_catch_compare(const char* throwtype, const char* catchtype, long* offset_result);
+extern char __throw_catch_compare(const char* throwtype, const char* catchtype, s32* offset_result);
 extern void __unexpected(CatchInfo* catchinfo);
 
 extern int __register_fragment(struct __eti_init_info* info, char* TOC);
